1.admin-增加导入结果记录实体类； 2.优化导入分析 3.thesis表增加grade

+++ rate/rateserver/rate-model/src/main/java/org/sys/rate/model/Thesis.java
@@ -50,7 +50,7 @@ public class Thesis {
     private String tutorNumber;
     private String studentNumber;
 
-    private Float grade;
+    private Double grade;
     private Integer startThesisId;
 
 }
